Use INT64_MAX instead of LLONG_MAX
These variables have the type int64_t, not long long. long long could
be a larger type than 64 bits. Emulate INT64_MAX for older versions of
MSVC, and remove the unreferenced vpx_ports/vpxtypes.h

+++ vp9/common/vp9_implicit_segmentation.c
@@ -33,8 +33,8 @@ typedef struct {
   int min_y;
   int max_x;
   int max_y;
-  long long sum_x;
-  long long sum_y;
+  int64_t sum_x;
+  int64_t sum_y;
   int pixels;
   int seg_value;
   int label;
